requirements.txt
Django==2.1.4
django-cors-headers==2.4.0
djangorestframework==3.9.0
pytz==2018.7
whitenoise==4.1.2
gunicorn==19.9.0